# services/sdkapi/docs/Sdk_swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "SDK API",
        "title": "Kanthor SDK API",
        "termsOfService": "http://kanthorlabs.com/terms/",
        "contact": {
            "name": "Kanthor Support",
            "url": "http://kanthorlabs.com/support",
            "email": "support@kanthorlabs.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/application": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "application"
                ],
                "parameters": [
                    {
                        "maxLength": 32,
                        "minLength": 29,
                        "type": "string",
                        "description": "current query cursor",
                        "name": "_cursor",
                        "in": "query"
                    },
                    {
                        "maxLength": 32,
                        "minLength": 2,
                        "type": "string",
                        "description": "search keyword",
                        "name": "_q",
                        "in": "query"
                    },
                    {
                        "maximum": 30,
                        "minimum": 5,
                        "type": "integer",
                        "description": "limit returning records",
                        "name": "_limit",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "only return records with selected ids",
                        "name": "_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.ApplicationListRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "application"
                ],
                "parameters": [
                    {
                        "description": "application properties",
                        "name": "props",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sdkapi.ApplicationCreateReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.ApplicationCreateRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            }
        },
...
        "/application/{app_id}/endpoint": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "endpoint"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "application id",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "maxLength": 32,
                        "minLength": 29,
                        "type": "string",
                        "description": "current query cursor",
                        "name": "_cursor",
                        "in": "query"
                    },
                    {
                        "maxLength": 32,
                        "minLength": 2,
                        "type": "string",
                        "description": "search keyword",
                        "name": "_q",
                        "in": "query"
                    },
                    {
                        "maximum": 30,
                        "minimum": 5,
                        "type": "integer",
                        "description": "limit returning records",
                        "name": "_limit",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "only return records with selected ids",
                        "name": "_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointListRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "endpoint"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "application id",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "endpoint properties",
                        "name": "props",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointCreateReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.EndpointCreateRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            }
        },
...
        "/application/{app_id}/message": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "message"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "application id",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "message properties",
                        "name": "props",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sdkapi.MessagePutReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/sdkapi.MessagePutRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/gateway.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.Application": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "I didn't find a way to disable automatic fields modify yet\nso, I use a tag to disable this feature here\nbut, we should keep our entities stateless if we can",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
        "entities.Endpoint": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "created_at": {
                    "description": "I didn't find a way to disable automatic fields modify yet\nso, I use a tag to disable this feature here\nbut, we should keep our entities stateless if we can",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "description": "HTTP: POST/PUT/PATCH",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "secret_key": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "uri": {
                    "description": "format: scheme \":\" [\"//\" authority] path [\"?\" query] [\"#\" fragment]\nHTTP: https:://httpbin.org/post?app=kanthor.webhook\ngRPC: grpc:://app.kanthorlabs.com",
                    "type": "string"
                }
            }
        },
...
        "gateway.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "sdkapi.ApplicationCreateReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "sdkapi.ApplicationCreateRes": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "I didn't find a way to disable automatic fields modify yet\nso, I use a tag to disable this feature here\nbut, we should keep our entities stateless if we can",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
...
        "sdkapi.ApplicationListRes": {
            "type": "object",
            "properties": {
                "cursor": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Application"
                    }
                }
            }
        },
...
        "sdkapi.EndpointCreateReq": {
            "type": "object",
            "required": [
                "method",
                "name",
                "uri"
            ],
            "properties": {
                "method": {
                    "type": "string",
                    "enum": [
                        "POST",
                        "PUT"
                    ]
                },
                "name": {
                    "type": "string"
                },
                "secret_key": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 16
                },
                "uri": {
                    "type": "string",
                    "example": "https://example.com"
                }
            }
        },
        "sdkapi.EndpointCreateRes": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "created_at": {
                    "description": "I didn't find a way to disable automatic fields modify yet\nso, I use a tag to disable this feature here\nbut, we should keep our entities stateless if we can",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "description": "HTTP: POST/PUT/PATCH",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "secret_key": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "uri": {
                    "description": "format: scheme \":\" [\"//\" authority] path [\"?\" query] [\"#\" fragment]\nHTTP: https:://httpbin.org/post?app=kanthor.webhook\ngRPC: grpc:://app.kanthorlabs.com",
                    "type": "string"
                }
            }
        },
...
        "sdkapi.EndpointListRes": {
            "type": "object",
            "properties": {
                "cursor": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Endpoint"
                    }
                }
            }
        },
...
        "sdkapi.MessagePutReq": {
            "type": "object",
            "required": [
                "body",
                "type"
            ],
            "properties": {
                "body": {
                    "type": "object",
                    "additionalProperties": true
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string",
                    "example": "testing.debug"
                }
            }
        },
        "sdkapi.MessagePutRes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}